refactor(web): migrate home from Pages to App Router
Move home page, navigation and sloth logo into app/, inline page
titles and drop the now-unused Pages Router scaffold, favicons and
shared barrel exports.
