+++ apps/web/app/sloth-face-square.png
