--- apps/web/components/home/index.ts
@@ -1,1 +1,0 @@
-export { default as Navigation } from './navigation';
