--- apps/web/constants/titles.ts
@@ -1,6 +1,0 @@
-const seriousSloth = 'SeriousSloth';
-const separator = '|';
-
-export const emotesTitle = `Emotes ${separator} ${seriousSloth}`;
-export const homeTitle = seriousSloth;
-
